removed stdint.h

+++ src/m0/c/m0_interp.c
@@ -71,9 +71,9 @@ new_call_frame( M0_Interp *interp ) {
     M0_CallFrame *frame = malloc( sizeof (M0_CallFrame) );
 
     /* this is a silly minimal hack for now */
-    frame->registers[CHUNK]  = (uint64_t)interp->first_chunk;
-    frame->registers[PC]     = (uint64_t)0;
-    frame->registers[CONSTS] = (uint64_t)interp->first_chunk->constants;
+    frame->registers[CHUNK]  = (UINT)interp->first_chunk;
+    frame->registers[PC]     = (UINT)0;
+    frame->registers[CONSTS] = (UINT)interp->first_chunk->constants;
 
     return frame;
 }
